app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.3
nltk==3.8.1
requests==2.31.0
gunicorn==21.2.0